weston: Add cleanup for .X*-lock files
This will make sure that the .X*-lock files are removed before
starting Weston. This is needed because the .X*-lock files are
created by the X server and if the container is restarted, the
files will still be there and the X server will create new ones
every time it is restarted.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>