scripts: pipeline: Fix PushToDockerhub parameter
For use it from another shell terminal, the parameter PushToDockerhub
should be a string, not a boolean. This way it is possible to pass the
value as a string from the command line. It should be "true" or "false".
Also we are getting the remaining arguments from the command line and
setting as the images paths to be built.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>