scripts: build: Fix unauthorized access to dockerhub
After long build process, the pipeline script was trying to push the
images to dockerhub with no success. Appears that the credentials were
somehow lost between the time from the docker login and the docker
compose push. So, I added a new call to the DockerRegistryLogin function
before the docker compose push command.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>